requirements.txt
gunicorn
flask
tensorflow
pillow 
numpy
